Abandon effort to sync the output. Can deal with it later.
I expected that it would work out of the box. Because:

1) https://cplusplus.com/reference/iostream/cout/ says
"By default, cout is synchronized with stdout "

2) https://cplusplus.com/reference/ostream/ostream/flush/ says
"Concurrent access to the same stream object may cause data races,
except for the standard stream objects (cout, cerr, clog) when these are
synchronized with stdio (in this case, no data races are initiated,
although no guarantees are given on the order in which characters
from multiple threads are inserted)."

So, I probably don't get some of the definitions used, or it's more
complicated. In either case, a problem for later.